+++ src/main/VideoProcessQueue.ts
@@ -376,8 +376,13 @@ export default class VideoProcessQueue {
       return;
     }
 
-    const sizeMonitor = new CloudSizeMonitor(this.mainWindow, this.cloudClient);
-    sizeMonitor.run();
+    const sizeMonitor = new CloudSizeMonitor(
+      this.mainWindow,
+      this.cloudClient,
+      250
+    );
+
+    await sizeMonitor.run();
     const usage = await sizeMonitor.usage();
 
     const status: CloudStatus = {